+++ project/example_configuration.gd
@@ -6,7 +6,7 @@ extends SentryConfiguration
 
 ## Configure Sentry SDK options
 func _configure(options: SentryOptions) -> void:
-	print("[example_configuration.gd] Configuring SDK options via GDScript")
+	print("INFO: [example_configuration.gd] Configuring SDK options via GDScript")
 
 	options.debug = true
 	options.release = "sentry-godot-demo@" + ProjectSettings.get_setting("application/config/version")
@@ -22,17 +22,17 @@ func _configure(options: SentryOptions) -> void:
 
 ## before_send callback example
 func _before_send(ev: SentryEvent) -> SentryEvent:
-	print("[example_configuration.gd] Processing event: ", ev.id)
+	print("INFO: [example_configuration.gd] Processing event: ", ev.id)
 	if ev.message.contains("Bruno"):
-		print("[example_configuration.gd] Removing sensitive information from the event")
+		print("INFO: [example_configuration.gd] Removing sensitive information from the event")
 		ev.message = ev.message.replace("Bruno", "REDACTED")
 	elif ev.message == "junk":
-		print("[example_configuration.gd] Discarding event with message 'junk'")
+		print("INFO: [example_configuration.gd] Discarding event with message 'junk'")
 		return null
 	return ev
 
 
 ## on_crash callback example
 func _on_crash(ev: SentryEvent) -> SentryEvent:
-	print("[example_configuration.gd] Crashing with event: ", ev.id)
+	print("INFO: [example_configuration.gd] Crashing with event: ", ev.id)
 	return ev